+++ Product/Finances/Domain/Repository/IOrcamentoRepository.cs
@@ -5,9 +5,9 @@ namespace Finances.Domain.Repository
     public interface IOrcamentoRepository : IFinancesCRUDRepository<Orcamento>
     {
         Orcamento GetOrcamentoPorVigencia(Vigencia vigencia);
-        OrcamentoCategoria GetOrcamentoCategoriaByID(int orcamentoCategoriaID);
-        void DeleteOrcamentoCategoria(int orcamentoCategoriaID);
-        int SaveOrcamentoCategoria(OrcamentoCategoria orcamentoCategoria);
-        void DeleteOrcamentoCategoriaPorCategoria(int categoriaID);
+        OrcamentoCategoria GetOrcamentoCategoriaByID(ulong orcamentoCategoriaID);
+        void DeleteOrcamentoCategoria(ulong orcamentoCategoriaID);
+        ulong SaveOrcamentoCategoria(OrcamentoCategoria orcamentoCategoria);
+        void DeleteOrcamentoCategoriaPorCategoria(ulong categoriaID);
     }
 }
